Move HubConversionFunction into functions folder
